rdu/rdu_data_logger/requirements.txt
aenum==3.1.0
pkg-resources==0.0.0
python-can==3.3.4
wrapt==1.12.1
